# docs/swagger.yaml
basePath: /
definitions:
  entities.HabitFrequency:
    enum:
    - diario
    - semanal
    type: string
    x-enum-varnames:
    - Daily
    - Weekly
  entities.HabitStreak:
    properties:
      createdAt:
        type: string
      currentStreak:
        type: integer
      frequency:
        $ref: '#/definitions/entities.HabitFrequency'
      habitName:
        type: string
      id:
        type: string
      maxStreak:
        type: integer
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  entities.Ritual:
    properties:
      agradecimiento:
        type: string
      createdAt:
        type: string
      date:
        type: string
      id:
        type: string
      metas:
        description: 'El éxito exige enfoque: exactamente 10 metas'
        items:
          type: string
        type: array
      userID:
        type: string
    type: object
  usecases.DailyRitualCommand:
    properties:
      Goals:
        example:
        - Finalizar la integración de pagos
        - Leer 20 páginas de un libro
        - Hacer 45 minutos de ejercicio
        items:
          type: string
        type: array
      GratefulFor:
        example: Estoy agradecido por la oportunidad de crear sistemas increíbles
          hoy
        type: string
      user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    type: object
info:
  contact: {}
  description: SOPRO_CORE_ENGINE
  title: SOPRO_CORE
  version: "0.1"
paths:
  /api/health:
    get:
      description: Verifica que el motor del sistema esté en línea y operando a su
        máxima capacidad.
      produces:
      - application/json
      responses:
        "200":
          description: Retorna un mensaje de victoria garantizada.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Comprueba la vitalidad del núcleo
      tags:
      - System
  /api/login/google:
    post:
      description: Gestiona el inicio de sesión o creación de cuenta. Requiere que
        el token de Firebase haya sido validado por el middleware.
      produces:
      - application/json
      responses:
        "200":
          description: Login exitoso y datos del usuario.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno o usuario no encontrado en el contexto.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autenticación maestra con Google
      tags:
      - Auht
  /api/rituals:
    get:
      description: Devuelve la lista histórica de todos los rituales diarios forjados
        por el usuario.
      produces:
      - application/json
      responses:
        "200":
          description: Historial obtenido con éxito.
          schema:
            items:
              $ref: '#/definitions/entities.Ritual'
            type: array
        "401":
          description: No autorizado.
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtiene el historial de rituales
      tags:
      - Ritual
    post:
      consumes:
      - application/json
      description: 'Registra el agradecimiento y las 10 metas innegociables del día.
        Regla inquebrantable: Solo se permite uno por día calendario.'
      parameters:
      - description: Tu enfoque y gratitud del día
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecases.DailyRitualCommand'
      produces:
      - application/json
      responses:
        "201":
          description: ¡Ritual completado! Tu enfoque está asegurado para hoy.
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Sin disciplina no hay resultados (No autorizado).
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Crea tu ritual diario
      tags:
      - Rituals
  /api/rituals/today:
    get:
      description: Retorna el agradecimiento y las metas establecidas para el día
        actual si ya fue creado.
      produces:
      - application/json
      responses:
        "200":
          description: Ritual del día obtenido con éxito.
          schema:
            type: boolean
      security:
      - BearerAuth: []
      summary: Verifica y obtiene el ritual de hoy
      tags:
      - Rituals
  /api/streaks/habit/{habitName}:
    get:
      consumes:
      - application/json
      description: Recupera la inercia actual y máxima de un hábito para medir tu
        nivel de disciplina.
      parameters:
      - description: Nombre del hábito (ej. Programacion)
        in: path
        name: habitName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: El estado actual de tu disciplina
          schema:
            $ref: '#/definitions/entities.HabitStreak'
      security:
      - BearerAuth: []
      summary: Obtener la racha de un hábito específico
      tags:
      - Streaks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"